page directories were done
