+++ app/routes/routes.ts
@@ -1,0 +1,6 @@
+export const routePaths = [
+  { href: "/", label: "Home" },
+  { href: "/pages/about-us", label: "About" },
+  { href: "/pages/products", label: "Products" },
+  { href: "/pages/contact-us", label: "Contact" },
+];
